# docs/api/openapi.locked.yaml
openapi: 3.1.0
info:
  title: FaultMaven API
  description: AI-powered troubleshooting assistant for Engineers, SREs, and DevOps
    professionals
  version: 1.0.0
paths:
  /api/v1/data:
    post:
      tags:
      - data_ingestion
      - data_processing
      summary: Upload Data Compat
      description: 'Compatibility endpoint for legacy tests - delegates to main upload
        function


        This endpoint maintains backward compatibility for existing tests that

        expect POST to /data instead of /data/upload.'
      operationId: upload_data_compat_api_v1_data_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_data_compat_api_v1_data_post'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/data/upload:
    post:
      tags:
      - data_ingestion
      - data_processing
      summary: Upload Data
      description: "Upload and process data with clean delegation pattern\n\nThis\
        \ endpoint follows the thin controller pattern:\n1. Basic input validation\
        \ (file size, type)\n2. Pure delegation to service layer for all business\
        \ logic\n3. Clean error boundary handling\n\nArgs:\n    file: File to upload\n\
        \    session_id: Session identifier \n    description: Optional description\
        \ of the data\n    data_service: Injected DataService from DI container\n\
        \    \nReturns:\n    UploadedData with processing results\n    \nRaises:\n\
        \    HTTPException: On service layer errors (400, 404, 413, 500)"
      operationId: upload_data_api_v1_data_upload_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_data_api_v1_data_upload_post'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/data/batch-upload:
    post:
      tags:
      - data_ingestion
      - data_processing
      summary: Batch Upload Data
      description: "Batch upload multiple files with clean delegation\n\nArgs:\n \
        \   files: List of files to upload\n    session_id: Session identifier\n \
        \   data_service: Injected DataService\n    \nReturns:\n    List of UploadedData\
        \ results"
      operationId: batch_upload_data_api_v1_data_batch_upload_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_batch_upload_data_api_v1_data_batch_upload_post'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/UploadedData'
                type: array
                title: Response Batch Upload Data Api V1 Data Batch Upload Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/data/{data_id}/analyze:
    post:
      tags:
      - data_ingestion
      - data_processing
      summary: Analyze Data
      description: "Analyze uploaded data with clean delegation\n\nArgs:\n    data_id:\
        \ Data identifier to analyze\n    analysis_request: Request containing session_id\
        \ and analysis parameters\n    data_service: Injected DataService\n    \n\
        Returns:\n    DataInsightsResponse with analysis results"
      operationId: analyze_data_api_v1_data__data_id__analyze_post
      parameters:
      - name: data_id
        in: path
        required: true
        schema:
          type: string
          title: Data Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: Analysis Request
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Analyze Data Api V1 Data  Data Id  Analyze Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/data/sessions/{session_id}:
    get:
      tags:
      - data_ingestion
      - data_processing
      summary: Get Session Data
      description: "Get all data for a session with clean delegation\n\nArgs:\n  \
        \  session_id: Session identifier\n    limit: Maximum number of results\n\
        \    offset: Pagination offset\n    data_service: Injected DataService\n \
        \   \nReturns:\n    List of UploadedData for the session"
      operationId: get_session_data_api_v1_data_sessions__session_id__get
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      - name: limit
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          default: 10
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Get Session Data Api V1 Data Sessions  Session Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/data/sessions/{session_id}/batch-process:
    post:
      tags:
      - data_ingestion
      - data_processing
      summary: Batch Process Session Data
      description: "Batch process data for a session\n\nArgs:\n    session_id: Session\
        \ identifier\n    batch_request: Request with data_ids to process\n    data_service:\
        \ Injected DataService\n    \nReturns:\n    Batch processing job information"
      operationId: batch_process_session_data_api_v1_data_sessions__session_id__batch_process_post
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: Batch Request
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Batch Process Session Data Api V1 Data Sessions  Session
                  Id  Batch Process Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/data/{data_id}:
    get:
      tags:
      - data_ingestion
      - data_processing
      summary: Get Data
      description: "Get data by ID\n\nArgs:\n    data_id: Data identifier\n    data_service:\
        \ Injected DataService\n    \nReturns:\n    Data information"
      operationId: get_data_api_v1_data__data_id__get
      parameters:
      - name: data_id
        in: path
        required: true
        schema:
          type: string
          title: Data Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadedData'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - data_ingestion
      - data_processing
      summary: Delete Data
      description: "Delete uploaded data with clean delegation\n\nArgs:\n    data_id:\
        \ Data identifier to delete\n    session_id: Session identifier for access\
        \ control (query parameter)\n    data_service: Injected DataService\n    \n\
        Returns:\n    Success confirmation"
      operationId: delete_data_api_v1_data__data_id__delete
      parameters:
      - name: data_id
        in: path
        required: true
        schema:
          type: string
          title: Data Id
      - name: session_id
        in: query
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/sessions/{session_id}:
    get:
      tags:
      - session_management
      - session_management
      summary: Get Session
      description: "Retrieve a specific session by ID.\n\nArgs:\n    session_id: Session\
        \ identifier\n\nReturns:\n    Session details"
      operationId: get_session_api_v1_sessions__session_id__get
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - session_management
      - session_management
      summary: Delete Session
      description: "Delete a specific session.\n\nArgs:\n    session_id: Session identifier\n\
        \nReturns:\n    Deletion confirmation"
      operationId: delete_session_api_v1_sessions__session_id__delete
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/cases/{case_id}/data:
    post:
      tags:
      - case_persistence
      - cases
      summary: Upload Case Data
      description: 'Upload data file to a specific case.


        Associates uploaded data with the case for context-aware troubleshooting.

        Returns 201 with Location header pointing to the created data resource.'
      operationId: upload_case_data_api_v1_cases__case_id__data_post
      parameters:
      - name: case_id
        in: path
        required: true
        schema:
          type: string
          title: Case Id
      - name: description
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Description of uploaded data
          title: Description
        description: Description of uploaded data
      - name: expected_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Expected data type
          title: Expected Type
        description: Expected data type
      responses:
        '201':
          description: Data uploaded successfully
          content:
            application/json:
              schema: {}
          headers:
            Location:
              description: URL of the created resource
              schema:
                type: string
            X-Correlation-ID:
              description: Request correlation ID
              schema:
                type: string
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - case_persistence
      - cases
      summary: List Case Data
      description: 'List data files associated with a case.


        Returns array of data records with pagination headers.

        Always returns 200 with empty array if no data exists.'
      operationId: list_case_data_api_v1_cases__case_id__data_get
      parameters:
      - name: case_id
        in: path
        required: true
        schema:
          type: string
          title: Case Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          description: Maximum number of items to return
          default: 50
          title: Limit
        description: Maximum number of items to return
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          description: Number of items to skip
          default: 0
          title: Offset
        description: Number of items to skip
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_batch_upload_data_api_v1_data_batch_upload_post:
      properties:
        files:
          items:
            type: string
            format: binary
          type: array
          title: Files
        session_id:
          type: string
          title: Session Id
      type: object
      required:
      - files
      - session_id
      title: Body_batch_upload_data_api_v1_data_batch_upload_post
    Body_upload_data_api_v1_data_upload_post:
      properties:
        file:
          type: string
          format: binary
          title: File
        session_id:
          type: string
          title: Session Id
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
      type: object
      required:
      - file
      - session_id
      title: Body_upload_data_api_v1_data_upload_post
    Body_upload_data_compat_api_v1_data_post:
      properties:
        file:
          type: string
          format: binary
          title: File
        session_id:
          type: string
          title: Session Id
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
      type: object
      required:
      - file
      - session_id
      title: Body_upload_data_compat_api_v1_data_post
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    UploadedData:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        type:
          $ref: '#/components/schemas/DataType'
        size_bytes:
          type: integer
          title: Size Bytes
        upload_timestamp:
          type: string
          title: Upload Timestamp
        processing_status:
          $ref: '#/components/schemas/ProcessingStatus'
        processing_summary:
          anyOf:
          - type: string
          - type: 'null'
          title: Processing Summary
        confidence_score:
          anyOf:
          - type: number
          - type: 'null'
          title: Confidence Score
      type: object
      required:
      - id
      - name
      - type
      - size_bytes
      - upload_timestamp
      - processing_status
      title: UploadedData
      description: A strongly-typed model for data uploaded by the user.
...